# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentWeather {
  /** Temperature in Celsius */
  temperature: number;
  /** Relative humidity in percent */
  humidity: number;
  /** Wind speed in km/h */
  windSpeed: number;
  /** WMO weather code */
  weatherCode: number;
  weatherDescription: string;
  /** Whether conditions are suitable for an open window */
  isWindowFriendly: boolean;
  /** Human-readable recommendation */
  recommendation: string;
  /** Reasons why window is or isn't recommended */
  reasons: string[];
  timestamp: string;
}

export interface HourlyWeather {
  hour: number;
  temperature: number;
  humidity: number;
  windSpeed: number;
  isWindowFriendly: boolean;
  weatherCode: number;
}

export interface WeatherForecast {
  date: string;
  hours: HourlyWeather[];
}

export interface TodaySummary {
  date: string;
  friendlyHoursCount: number;
  /** Hour of day (0-23) when best window-open period begins */
  bestWindowStart: number | null;
  bestWindowEnd: number | null;
  minTemp: number;
  maxTemp: number;
  overallRecommendation: string;
}

export interface Settings {
  id: number;
  /** Minimum comfortable temperature (Celsius) */
  minTemp: number;
  /** Maximum comfortable temperature (Celsius) */
  maxTemp: number;
  /** Maximum acceptable humidity (%) */
  maxHumidity: number;
  /** Maximum acceptable wind speed (km/h) */
  maxWindSpeed: number;
  /** Work days as integers: 0=Sun, 1=Mon, ..., 6=Sat */
  workDays: number[];
  /** Work start hour (0-23) */
  workStartHour: number;
  /** Work end hour (0-23) */
  workEndHour: number;
  notificationsEnabled: boolean;
  /** How often to check weather in minutes */
  checkIntervalMinutes: number;
}

export interface SettingsUpdate {
  minTemp?: number;
  maxTemp?: number;
  maxHumidity?: number;
  maxWindSpeed?: number;
  workDays?: number[];
  workStartHour?: number;
  workEndHour?: number;
  notificationsEnabled?: boolean;
  checkIntervalMinutes?: number;
}

export type WindowEventAction =
  (typeof WindowEventAction)[keyof typeof WindowEventAction];

export const WindowEventAction = {
  opened: "opened",
  closed: "closed",
} as const;

/**
 * Whether event was manual or triggered by weather check
 */
export type WindowEventTriggeredBy =
  (typeof WindowEventTriggeredBy)[keyof typeof WindowEventTriggeredBy];

export const WindowEventTriggeredBy = {
  user: "user",
  auto: "auto",
} as const;

export interface WindowEvent {
  id: number;
  action: WindowEventAction;
  /** Whether event was manual or triggered by weather check */
  triggeredBy: WindowEventTriggeredBy;
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  note?: string | null;
  createdAt: string;
}

export type CreateWindowEventAction =
  (typeof CreateWindowEventAction)[keyof typeof CreateWindowEventAction];

export const CreateWindowEventAction = {
  opened: "opened",
  closed: "closed",
} as const;

export type CreateWindowEventTriggeredBy =
  (typeof CreateWindowEventTriggeredBy)[keyof typeof CreateWindowEventTriggeredBy];

export const CreateWindowEventTriggeredBy = {
  user: "user",
  auto: "auto",
} as const;

export interface CreateWindowEvent {
  action: CreateWindowEventAction;
  triggeredBy: CreateWindowEventTriggeredBy;
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  note?: string | null;
}

export interface EventStats {
  totalOpens: number;
  totalCloses: number;
  avgOpenDurationMinutes: number | null;
  mostCommonOpenHour: number | null;
  daysWithWindowOpen: number;
}

export type GetCurrentWeatherParams = {
  lat: number;
  lon: number;
};

export type GetWeatherForecastParams = {
  lat: number;
  lon: number;
};

export type GetTodaySummaryParams = {
  lat: number;
  lon: number;
};

export type GetEventsParams = {
  limit?: number;
};