# lib/api-zod/src/generated/types/createWindowEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWindowEventAction } from "./createWindowEventAction";
import type { CreateWindowEventTriggeredBy } from "./createWindowEventTriggeredBy";

export interface CreateWindowEvent {
  action: CreateWindowEventAction;
  triggeredBy: CreateWindowEventTriggeredBy;
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  note?: string | null;
}