# lib/api-zod/src/generated/types/eventStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */

export interface EventStats {
  totalOpens: number;
  totalCloses: number;
  avgOpenDurationMinutes: number | null;
  mostCommonOpenHour: number | null;
  daysWithWindowOpen: number;
}