# lib/api-zod/src/generated/types/hourlyWeather.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */

export interface HourlyWeather {
  hour: number;
  temperature: number;
  humidity: number;
  windSpeed: number;
  isWindowFriendly: boolean;
  weatherCode: number;
}