# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  minTemp?: number;
  maxTemp?: number;
  maxHumidity?: number;
  maxWindSpeed?: number;
  workDays?: number[];
  workStartHour?: number;
  workEndHour?: number;
  notificationsEnabled?: boolean;
  checkIntervalMinutes?: number;
}