# lib/api-zod/src/generated/types/todaySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */

export interface TodaySummary {
  date: string;
  friendlyHoursCount: number;
  /** Hour of day (0-23) when best window-open period begins */
  bestWindowStart: number | null;
  bestWindowEnd: number | null;
  minTemp: number;
  maxTemp: number;
  overallRecommendation: string;
}