# lib/api-zod/src/generated/types/windowEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Window Weather Notifier API
 * OpenAPI spec version: 0.1.0
 */
import type { WindowEventAction } from "./windowEventAction";
import type { WindowEventTriggeredBy } from "./windowEventTriggeredBy";

export interface WindowEvent {
  id: number;
  action: WindowEventAction;
  /** Whether event was manual or triggered by weather check */
  triggeredBy: WindowEventTriggeredBy;
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  note?: string | null;
  createdAt: Date;
}